Update for TF-1.4.0
This guide was very helpful for building TensorFlow on my new RPi3, but I found several differences when working with TensorFlow-1.4.0.  Mostly these are obvious, but finding a working Eigen took a few tries, so I wanted to feed back that fix (although I assume it will get updated upstream in due course).